refactor environments rules to the separate namespace
