+++ src/python/pants/backend/adhoc/adhoc_tool.py
@@ -26,6 +26,11 @@ from pants.backend.adhoc.target_types import (
     AdhocToolWorkspaceInvalidationSourcesField,
 )
 from pants.base.glob_match_error_behavior import GlobMatchErrorBehavior
+from pants.core.environments.rules import (
+    EnvironmentNameRequest,
+    get_target_for_environment_name,
+    resolve_environment_name,
+)
 from pants.core.target_types import FileSourceField
 from pants.core.util_rules.adhoc_process_support import (
     AdhocProcessRequest,
@@ -35,11 +40,6 @@ from pants.core.util_rules.adhoc_process_support import (
     prepare_env_vars,
 )
 from pants.core.util_rules.adhoc_process_support import rules as adhoc_process_support_rules
-from pants.core.util_rules.environments import (
-    EnvironmentNameRequest,
-    get_target_for_environment_name,
-    resolve_environment_name,
-)
 from pants.engine.environment import EnvironmentName
 from pants.engine.fs import PathGlobs
 from pants.engine.intrinsics import digest_to_snapshot
