+++ src/python/pants/bsp/protocol.py
@@ -19,7 +19,7 @@ from pylsp_jsonrpc.streams import (  # type: ignore[import-untyped]
 
 from pants.bsp.context import BSPContext
 from pants.bsp.spec.notification import BSPNotification
-from pants.core.util_rules.environments import determine_bootstrap_environment
+from pants.core.environments.rules import determine_bootstrap_environment
 from pants.engine.environment import EnvironmentName
 from pants.engine.fs import Workspace
 from pants.engine.internals.scheduler import SchedulerSession
