+++ src/python/pants/core/environments/rules_integration_test.py
