+++ src/python/pants/core/goals/run.py
@@ -10,8 +10,8 @@ from dataclasses import dataclass
 from enum import Enum
 from typing import Any, ClassVar, TypeVar, final
 
+from pants.core.environments.rules import _warn_on_non_local_environments
 from pants.core.subsystems.debug_adapter import DebugAdapterSubsystem
-from pants.core.util_rules.environments import _warn_on_non_local_environments
 from pants.engine.env_vars import CompleteEnvironmentVars
 from pants.engine.environment import EnvironmentName
 from pants.engine.fs import Digest, Workspace
