+++ src/python/pants/goal/help.py
@@ -9,7 +9,7 @@ from typing import ClassVar
 from pants.base.exiter import ExitCode
 from pants.base.specs import Specs
 from pants.build_graph.build_configuration import BuildConfiguration
-from pants.core.util_rules.environments import determine_bootstrap_environment
+from pants.core.environments.rules import determine_bootstrap_environment
 from pants.engine.internals.parser import BuildFileSymbolsInfo
 from pants.engine.internals.selectors import Params
 from pants.engine.target import RegisteredTargetTypes
