+++ src/python/pants/init/engine_initializer.py
@@ -15,8 +15,9 @@ from pants.base.build_root import BuildRoot
 from pants.base.exiter import PANTS_SUCCEEDED_EXIT_CODE
 from pants.base.specs import Specs
 from pants.build_graph.build_configuration import BuildConfiguration
-from pants.core.util_rules import environments, system_binaries
-from pants.core.util_rules.environments import determine_bootstrap_environment
+from pants.core.environments import rules as environments_rules
+from pants.core.environments.rules import determine_bootstrap_environment
+from pants.core.util_rules import system_binaries
 from pants.engine import desktop, download_file, fs, intrinsics, process
 from pants.engine.console import Console
 from pants.engine.environment import EnvironmentName
@@ -297,7 +298,7 @@ class EngineInitializer:
                 *specs_rules.rules(),
                 *options_parsing.rules(),
                 *process.rules(),
-                *environments.rules(),
+                *environments_rules.rules(),
                 *system_binaries.rules(),
                 *platform_rules.rules(),
                 *changed_rules(),
