+++ src/python/pants/option/subsystem.py
@@ -104,7 +104,7 @@ class Subsystem(metaclass=_SubsystemMeta):
         _options_env: EnvironmentVars = EnvironmentVars()
 
         def __getattribute__(self, __name: str) -> Any:
-            from pants.core.util_rules.environments import resolve_environment_sensitive_option
+            from pants.core.environments.rules import resolve_environment_sensitive_option
 
             # Will raise an `AttributeError` if the attribute is not defined.
             # MyPy should stop that from ever happening.
@@ -132,7 +132,7 @@ class Subsystem(metaclass=_SubsystemMeta):
 
         def _is_default(self, __name: str) -> bool:
             """Returns true if the value of the named option is unchanged from the default."""
-            from pants.core.util_rules.environments import resolve_environment_sensitive_option
+            from pants.core.environments.rules import resolve_environment_sensitive_option
 
             v = getattr(type(self), __name)
             assert isinstance(v, OptionInfo)
@@ -147,7 +147,7 @@ class Subsystem(metaclass=_SubsystemMeta):
     def rules(cls: Any) -> Iterable[Rule]:
         # NB: This avoids using `memoized_classmethod` until its interaction with `mypy` can be improved.
         if cls._rules is None:
-            from pants.core.util_rules.environments import add_option_fields_for
+            from pants.core.environments.rules import add_option_fields_for
             from pants.engine.rules import Rule
 
             # nb. `rules` needs to be memoized so that repeated calls to add these rules
